+++ beeref/gui.py
@@ -16,33 +16,111 @@
 import logging
 import os.path
 
-from PyQt6 import QtCore, QtWidgets
+from PyQt6 import QtCore, QtGui, QtWidgets
 from PyQt6.QtCore import Qt
 
 from beeref import constants
-from beeref.config import logfile_name
+from beeref.config import logfile_name, BeeSettings
+from beeref.main_controls import MainControlsMixin
 
 
 logger = logging.getLogger(__name__)
 
 
-class WelcomeOverlay(QtWidgets.QWidget):
+class RecentFilesModel(QtCore.QAbstractListModel):
+    """An entry in the 'Recent Files' list."""
+
+    def __init__(self, files):
+        super().__init__()
+        self.files = files
+
+    def rowCount(self, parent):
+        return len(self.files)
+
+    def data(self, index, role):
+        if role == QtCore.Qt.ItemDataRole.DisplayRole:
+            return os.path.basename(self.files[index.row()])
+        if role == QtCore.Qt.ItemDataRole.FontRole:
+            font = QtGui.QFont()
+            font.setUnderline(True)
+            return font
+
+
+class RecentFilesView(QtWidgets.QListView):
+
+    def __init__(self, parent, files=None):
+        super().__init__(parent)
+        self.files = files or []
+        self.clicked.connect(self.on_clicked)
+        self.setModel(RecentFilesModel(self.files))
+        self.setMouseTracking(True)
+
+    def on_clicked(self, index):
+        self.parent().parent().open_from_file(self.files[index.row()])
+
+    def update_files(self, files):
+        self.files = files
+        self.model().files = files
+        self.reset()
+
+    def sizeHint(self):
+        size = QtCore.QSize()
+        height = sum(
+            (self.sizeHintForRow(i) + 2) for i in range(len(self.files)))
+        width = max(self.sizeHintForColumn(i) for i in range(len(self.files)))
+        size.setHeight(height)
+        size.setWidth(width + 2)
+        return size
+
+    def mouseMoveEvent(self, event):
+        index = self.indexAt(
+            QtCore.QPoint(int(event.position().x()),
+                          int(event.position().y())))
+        if index.isValid():
+            self.setCursor(Qt.CursorShape.PointingHandCursor)
+        else:
+            self.setCursor(Qt.CursorShape.ArrowCursor)
+
+        super().mouseMoveEvent(event)
+
+
+class WelcomeOverlay(MainControlsMixin, QtWidgets.QWidget):
     """Some basic info to be displayed when the scene is empty."""
 
     txt = """<p>Paste or drop images here.</p>
              <p>Right-click for more options.</p>"""
 
-    def __init__(self, parent=None):
+    def __init__(self, parent):
         super().__init__(parent)
+        self.control_target = parent
         self.setAttribute(Qt.WidgetAttribute.WA_NoSystemBackground)
-        self.setAttribute(Qt.WidgetAttribute.WA_TransparentForMouseEvents)
-        label = QtWidgets.QLabel(self)
-        label.setText(self.txt)
+        self.init_main_controls()
+
+        # Recent files
+        self.files_layout = QtWidgets.QVBoxLayout()
+        self.files_layout.addStretch(50)
+        self.files_layout.addWidget(
+            QtWidgets.QLabel('<h3>Recent Files</h3>', self))
+        self.files_view = RecentFilesView(self)
+        self.files_layout.addWidget(self.files_view)
+        self.files_layout.addStretch(50)
+
+        # Help text
+        label = QtWidgets.QLabel(self.txt, self)
         label.setAlignment(Qt.AlignmentFlag.AlignVCenter
                            | Qt.AlignmentFlag.AlignCenter)
-        layout = QtWidgets.QVBoxLayout()
-        layout.addWidget(label)
-        self.setLayout(layout)
+        self.layout = QtWidgets.QHBoxLayout()
+        self.layout.addStretch(50)
+        self.layout.addWidget(label)
+        self.layout.addStretch(50)
+        self.setLayout(self.layout)
+
+    def show(self):
+        files = BeeSettings().get_recent_files(existing_only=True)
+        self.files_view.update_files(files)
+        if files and self.layout.indexOf(self.files_layout) < 0:
+            self.layout.insertLayout(0, self.files_layout)
+        super().show()
 
 
 class BeeProgressDialog(QtWidgets.QProgressDialog):