+++ beeref/main_controls.py
@@ -1,0 +1,78 @@
+# This file is part of BeeRef.
+#
+# BeeRef is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# BeeRef is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with BeeRef.  If not, see <https://www.gnu.org/licenses/>.
+
+import logging
+
+from PyQt6 import QtCore, QtGui
+from PyQt6.QtCore import Qt
+
+from beeref import commands
+from beeref.items import BeePixmapItem
+from beeref import fileio
+
+
+logger = logging.getLogger(__name__)
+
+
+class MainControlsMixin:
+    """Basic controls shared by the main view and the welcome overlay:
+
+    * Right-click menu
+    * Dropping files
+    """
+
+    def init_main_controls(self):
+        self.setContextMenuPolicy(
+            Qt.ContextMenuPolicy.CustomContextMenu)
+        self.customContextMenuRequested.connect(
+            self.control_target.on_context_menu)
+        self.setAcceptDrops(True)
+
+    def dragEnterEvent(self, event):
+        mimedata = event.mimeData()
+        logger.debug(f'Drag enter event: {mimedata.formats()}')
+        if mimedata.hasUrls():
+            event.acceptProposedAction()
+        elif mimedata.hasImage():
+            event.acceptProposedAction()
+        else:
+            logger.info('Attempted drop not an image')
+
+    def dragMoveEvent(self, event):
+        event.acceptProposedAction()
+
+    def dropEvent(self, event):
+        mimedata = event.mimeData()
+        logger.debug(f'Handling file drop: {mimedata.formats()}')
+        pos = QtCore.QPoint(round(event.position().x()),
+                            round(event.position().y()))
+        if mimedata.hasUrls():
+            logger.debug(f'Found dropped urls: {mimedata.urls()}')
+            if not self.control_target.scene.items():
+                # Check if we have a bee file we can open directly
+                path = mimedata.urls()[0]
+                if (path.isLocalFile()
+                        and fileio.is_bee_file(path.toLocalFile())):
+                    self.control_target.open_from_file(path.toLocalFile())
+                    return
+            self.control_target.do_insert_images(mimedata.urls(), pos)
+        elif mimedata.hasImage():
+            img = QtGui.QImage(mimedata.imageData())
+            item = BeePixmapItem(img)
+            pos = self.control_target.mapToScene(pos)
+            self.control_target.undo_stack.push(
+                commands.InsertItems(self.control_target.scene, [item], pos))
+        else:
+            logger.info('Drop not an image')