+++ tests/test_gui.py
@@ -1,8 +1,10 @@
-from PyQt6 import QtWidgets
+from unittest.mock import MagicMock
+
+from PyQt6 import QtCore, QtWidgets
 from PyQt6.QtCore import Qt
 
 from beeref.config import logfile_name
-from beeref.gui import DebugLogDialog
+from beeref.gui import DebugLogDialog, RecentFilesModel
 
 
 def test_debug_log_dialog(qtbot, settings, view):
@@ -16,3 +18,23 @@ def test_debug_log_dialog(qtbot, settings, view):
     qtbot.mouseClick(dialog.copy_button, Qt.MouseButton.LeftButton)
     clipboard = QtWidgets.QApplication.clipboard()
     assert clipboard.text() == 'my log output'
+
+
+def test_recent_files_model_rowcount(view):
+    model = RecentFilesModel(['foo.png', 'bar.png'])
+    assert model.rowCount(None) == 2
+
+
+def test_recent_files_model_data_diplayrole(view):
+    model = RecentFilesModel(['foo.png', 'bar.png'])
+    index = MagicMock()
+    index.row.return_value = 1
+    assert model.data(index, QtCore.Qt.ItemDataRole.DisplayRole) == 'bar.png'
+
+
+def test_recent_files_model_data_fontrole(view):
+    model = RecentFilesModel(['foo.png', 'bar.png'])
+    index = MagicMock()
+    index.row.return_value = 1
+    font = model.data(index, QtCore.Qt.ItemDataRole.FontRole)
+    assert font.underline() is True